重构wpcos插件，1. 修改数据结构，2.优化了上传钩子，3.优化了删除功能，减少请求次数，4.wp插件提交规范优化，5.升级时数据结构一键转化

+++ wpcos/uninstall.php
@@ -1,16 +1,20 @@
 <?php
 /**
- * Created by PhpStorm.
+* Created by PhpStorm.
  * User: zdl25
- * Date: 2019/1/21
- * Time: 10:50
- */
+         * Date: 2019/1/21
+                 * Time: 10:50
+                            */
 if(!defined('WP_UNINSTALL_PLUGIN')){
 	// 如果 uninstall 不是从 WordPress 调用，则退出
 	exit();
 }
 
 // 从 options 表删除选项
-delete_option( 'xos_options' );
+if (get_option('xos_options')) {
+	delete_option( 'xos_options' );
+}
+delete_option('wpcos_options');
 
 // 删除其他额外的选项和自定义表
+update_option('upload_url_path', '');
